refactor: extract star button to separate module
Extract star button component from mobile.ts into dedicated starButton.ts module to improve code organization and reduce file size.

## Changes

**New File: `src/components/starButton.ts`**
- Exported `createStarButton()` function with config-based API
- Exported `StarButtonConfig` and `StarButton` interfaces
- Self-contained icon SVG constants (filled and outline stars)
- Encapsulated button appearance update logic

**Updated: `src/components/mobile.ts`**
- Import `createStarButton` from new module
- Remove inline star button implementation (~70 lines)
- Update star button creation calls to use config object:
  - `getIsStarred: () => boolean`
  - `onToggle: (isStarred: boolean) => void`
  - `onResetFade?: () => void`
  - `isTransitioning: () => boolean`

## Benefits

- **Modularity**: Star button logic isolated in dedicated file
- **Reusability**: Can be imported and used in other contexts
- **Maintainability**: Easier to locate and modify star button behavior
- **File Size**: Reduces mobile.ts by ~70 lines
- **Consistency**: Follows same pattern as other extracted components

## Testing

- ✅ TypeScript compilation passes
- ✅ Linter passes
- ✅ Build succeeds
- ✅ No functional changes to behavior
